Third attempt at solution, got it working.
My second attempt involved iteratively removing the smallest element from the list until we hit `k` elements but that didn't work since it doesn't guarantee that the list is lexically the largest. Consider the example: "234234234234278". Blindly removing the smallest element (2 in this case), will give the number "3434343...". But the example clearly lays out "434234234278", which DOES contain 2s. So this logic is flawed.

What DOES work is this approach: we have to find the lexically largest sublist of length `len()-1` iteratively and it works. I considered brute-forcing the results, but with the help f ChatGPT realized that it isn't necessary since lexical comparison is essentially finding the FIRST time they differ.